+++ src/Marvin.Migrations/MigratorBuilder.cs
@@ -23,6 +23,14 @@ namespace Marvin.Migrations
 
         private ILogger _logger;
 
+        /// <summary>
+        /// Dictionary with variables
+        /// </summary>
+        /// <remarks>
+        /// Key = variable name, value - variable value
+        /// </remarks>
+        private readonly Dictionary<string, string> _variables;
+
         private readonly IServiceCollection _services;
 
         /// <inheritdoc />
@@ -34,6 +42,7 @@ namespace Marvin.Migrations
             _upgradePolicy = MigrationPolicy.All;
             _downgradePolicy = MigrationPolicy.All;
             _targetVersion = default(DbVersion?);
+            _variables = new Dictionary<string, string>();
         }
         
         /// <summary>
@@ -153,6 +162,23 @@ namespace Marvin.Migrations
         }
 
         /// <summary>
+        /// Add specified variables that will be passed to code migrations and auto substitute to script migrations
+        /// </summary>
+        /// <param name="name">Variable name</param>
+        /// <param name="value">Variable value</param>
+        /// <returns></returns>
+        /// <exception cref="ArgumentNullException">If any of arguments is <see langword="null"/> or empty</exception>
+        public MigratorBuilder UseVariable(string name, string value)
+        {
+            if (String.IsNullOrWhiteSpace(name))
+                throw new ArgumentNullException(nameof(name));
+
+            _variables[name] = value ?? throw new ArgumentNullException(nameof(value));
+
+            return this;
+        }
+
+        /// <summary>
         /// Build migrator
         /// </summary>
         /// <returns></returns>
@@ -164,23 +190,32 @@ namespace Marvin.Migrations
             
             var dbProvider = _dbProviderFactory.CreateDbProvider();
             
+            var providerVariables = dbProvider.GetDefaultVariables();
+            foreach (var kvp in providerVariables)
+            {
+                // we should not override the variables set manually
+                if (_variables.ContainsKey(kvp.Key)) continue;
+
+                _variables[kvp.Key] = kvp.Value;
+            }
+            
             var migrations = new List<IMigration>();
             foreach (var migrationsProvider in _migrationsProviders)
             {
-                migrations.AddRange(migrationsProvider.GetMigrations(dbProvider));
+                migrations.AddRange(migrationsProvider.GetMigrations(dbProvider, _variables));
             }
             
             var preMigrations = new List<IMigration>();
             foreach (var migrationsProvider in _preMigrationsProviders)
             {
-                preMigrations.AddRange(migrationsProvider.GetMigrations(dbProvider));
+                preMigrations.AddRange(migrationsProvider.GetMigrations(dbProvider, _variables));
             }
             
             return new DbMigrator(
                 dbProvider, 
                 migrations, 
                 _upgradePolicy, 
-                _downgradePolicy, 
+                _downgradePolicy,
                 preMigrations,
                 _targetVersion, 
                 _logger);