+++ tests/IntegrationTests/Marvin.Migrations.TransactionTests/CodeMigrations/CodeMigration_3_0.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Data.Common;
 using System.Threading.Tasks;