+++ tests/UnitTests/Marvin.Migrations.UnitTests/CodeMigrations/CustomBaseCodeMigration.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace Marvin.Migrations.UnitTests.CodeMigrations
 {
     public abstract class CustomBaseCodeMigration : CodeMigration