+++ tests/UnitTests/Marvin.Migrations.UnitTests/DbMigratorTests.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using System.Transactions;
 using Moq;
 using Xunit;
+using IsolationLevel = System.Data.IsolationLevel;
 
 namespace Marvin.Migrations.UnitTests
 {
@@ -21,6 +22,10 @@ namespace Marvin.Migrations.UnitTests
             provider
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
+
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
 
             var migrations = new List<IMigration>(0);
             
@@ -70,6 +75,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.OpenConnectionAsync())
                 .Callback(() => {})
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -123,6 +132,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -166,6 +179,10 @@ namespace Marvin.Migrations.UnitTests
             
             var provider = new Mock<IDbProvider>();
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             provider
                 .Setup(x => x.UpdateCurrentDbVersionAsync(It.IsAny<DbVersion>()))
                 .Callback<DbVersion>(version => dbVersionAfterUpdate = version)
@@ -236,6 +253,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -290,6 +311,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -350,6 +375,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -407,6 +436,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -468,6 +501,10 @@ namespace Marvin.Migrations.UnitTests
                 .Setup(x => x.GetDbVersionSafeAsync())
                 .Returns(() => Task.FromResult(new DbVersion?(initialDbVersion)));
 
+            provider
+                .Setup(x => x.BeginTransaction())
+                .Returns(() => new MockTransaction());
+            
             var firstMigration = new Mock<IMigration>();
             firstMigration
                 .Setup(x => x.Version)
@@ -508,5 +545,20 @@ namespace Marvin.Migrations.UnitTests
             Assert.Equal(MigrationError.PolicyError, result.Error.Value);
             Assert.Equal(thirdMigration.Object.Version, dbVersionAfterUpdate);
         }
+        
+        private class MockTransaction : DbTransaction
+        {
+            public override void Commit()
+            {
+            }
+
+            public override void Rollback()
+            {
+            }
+
+            protected override DbConnection DbConnection { get; }
+            public override IsolationLevel IsolationLevel { get; }
+        }
     }
+    
 }