+++ src/Marvin.Migrations/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.3.1] - 2019-09-16
+
+### Fixed 
+
+- Passing variables to code migrations
+
+## [0.3] - 2019-09-16
+
+### Added 
+
+- Variables for scripts with auto substitution to script migrations 
+
 ## [0.2.16] - 2018-12-19
 
 ### Added